% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo2_compendium-package.R
\docType{package}
\name{demo2_compendium-package}
\alias{demo2_compendium}
\alias{demo2_compendium-package}
\title{demo2_compendium: The Title of the Project}
\description{
A paragraph providing a full description of the project (on several lines...)
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jacquetclaire/demo2_compendium}
  \item Report bugs at \url{https://github.com/jacquetclaire/demo2_compendium/issues}
}

}
\author{
\strong{Maintainer}: Claire Jacquet \email{claire.jacquet@umontpellier.fr} (\href{https://orcid.org/0000-0002-7913-0689}{ORCID}) [copyright holder]

}
\keyword{internal}
